Update pagination to work with filtering
